pyproject.toml
[project]
name = "halolight-api-python"
version = "1.0.0"
description = "HaloLight Backend API - Python/FastAPI/SQLAlchemy implementation"
authors = [{name = "HaloLight Team"}]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["halolight", "api", "fastapi", "sqlalchemy", "python"]
license = "ISC"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.19",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.1.0",      # Email validation for Pydantic EmailStr
    "python-dateutil>=2.9.0",      # Date utilities for calendar module
    "aiofiles>=24.1.0",            # Async file operations for file uploads
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "UP037", # Remove quotes from type annotation (needed for TYPE_CHECKING)
]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
# Relaxed settings for initial development
# TODO: Enable strict mode after fixing all type errors
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = false
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
